Rename FarmList to ProductionStatusByFarm.
